fix: secure /api/chat endpoint with authentication
