+++ client/src/modules/ai-assistant/components/ChatBox.jsx
@@ -1,7 +1,10 @@
 import { useState } from "react";
+import { useSelector } from "react-redux";
+import { apiRequest } from "../../../services/apiClient";
 import MessageBubble from "./MessageBubble";
 
 const ChatBox = () => {
+  const { token } = useSelector((state) => state.auth);
   const [messages, setMessages] = useState([
     { sender: "bot", text: "Hi! How can I help you?" },
   ]);
@@ -11,15 +14,11 @@ const ChatBox = () => {
   // 🔥 NEW: backend call function
   const sendMessageToBackend = async (message) => {
     try {
-      const res = await fetch("http://localhost:5000/api/chat", {
+      const data = await apiRequest("/api/chat", {
         method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ message }),
+        token,
+        body: { message },
       });
-
-      const data = await res.json();
       return data.reply;
     } catch (error) {
       console.error(error);
